+++ yazi-config/src/which/sorting.rs
@@ -30,14 +30,3 @@ impl TryFrom<String> for SortBy {
 
 	fn try_from(value: String) -> Result<Self, Self::Error> { Self::from_str(&value) }
 }
-
-impl ToString for SortBy {
-	fn to_string(&self) -> String {
-		match self {
-			Self::None => "none",
-			Self::Key => "key",
-			Self::Desc => "desc",
-		}
-		.to_string()
-	}
-}
