+++ yazi-core/src/which/mod.rs
@@ -1,8 +1,6 @@
 mod commands;
-mod config;
 mod sorter;
 mod which;
 
-pub use config::*;
 pub use sorter::*;
 pub use which::*;
